Change decimals to be of type uint8 instead of uint256 per ERC20 standard
OpenZeppelin Audit Low Severity Recommendation:
According to the ERC20 standard the decimals variable (defined in line 31) should be of type uint8.

See:
https://github.com/frozeman/EIPs/blob/94bc4311e889c2c58c561c074be1483f48ac9374/EIPS/eip-20-token-standard.md#decimals

+++ contracts/TierionNetworkToken.sol
@@ -26,10 +26,10 @@ import "./zeppelin-solidity/contracts/lifecycle/Pausable.sol";
  */
 contract TierionNetworkToken is StandardToken, Pausable {
 
-  string public constant name = 'Tierion Network Token';                 // Set the token name for display
-  string public constant symbol = 'TNT';                                 // Set the token symbol for display
-  uint256 public constant decimals = 8;                                  // Set the number of decimals for display
-  uint256 public constant INITIAL_SUPPLY = 1000000000 * 10**decimals;    // 1 Billion TNT specified in Grains
+  string public constant name = 'Tierion Network Token';                       // Set the token name for display
+  string public constant symbol = 'TNT';                                       // Set the token symbol for display
+  uint8 public constant decimals = 8;                                          // Set the number of decimals for display
+  uint256 public constant INITIAL_SUPPLY = 1000000000 * 10**uint256(decimals); // 1 Billion TNT specified in Grains
 
   /**
    * @dev TierionNetworkToken Constructor
